[Fast Refresh] Separate LogBox Tests
